requirements.txt
streamlit
langchain
langchain-openai
langchain-core
chromadb
pypdf
python-dotenv
tiktoken
